Functional core, imperative shell (#77)
