+++ crates/sage-cli/tests/integration.rs
@@ -8,16 +8,17 @@ use sage_core::spectrum::SpectrumProcessor;
 #[test]
 fn integration() -> anyhow::Result<()> {
     let mut builder = Builder::default();
-    builder.update_fasta("../../tests/Q99536.fasta".into());
+    builder.update_fasta("foo".into());
 
-    let database = builder.make_parameters().build()?;
+    let fasta = sage_cloudpath::util::read_fasta("../../tests/Q99536.fasta", "rev_", true)?;
+    let database = builder.make_parameters().build(fasta);
     let pep = Peptide::try_from(Digest {
         sequence: "LQSRPAAPPAPGPGQLTLR".into(),
         ..Default::default()
     })
     .unwrap();
 
-    let spectra = sage_core::read_mzml("../../tests/LQSRPAAPPAPGPGQLTLR.mzML", None)?;
+    let spectra = sage_cloudpath::util::read_mzml("../../tests/LQSRPAAPPAPGPGQLTLR.mzML", None)?;
     assert_eq!(spectra.len(), 1);
 
     let sp = SpectrumProcessor::new(100, 0.0, 1500.0, true, 0);
