+++ crates/sage-cloudpath/Cargo.toml
@@ -11,11 +11,20 @@ license = "MIT"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-async-compression = { version = "0.3", features = ["tokio", "gzip"] }
+anyhow = "1.0"
+async-compression = { version = "0.3", features = ["tokio", "gzip", "zlib"] }
 aws-config = "0.54.1"
 aws-sdk-s3 = "0.24"
+base64 = "0.13"
 bytes = "1.0"
 http = "0.2"
 futures = "0.3"
+log = "0.4.0"
 once_cell = "1.0"
-tokio = { version = "1.0", features = ["fs", "io-util", "rt"] }
+tokio = { version = "1.0", features = ["fs", "io-util", "rt", "macros"] }
+quick-xml = { version = "0.25", features = ["async-tokio"] }
+
+serde = { version="1.0", features = ["derive"] }
+serde_json = "1.0"
+
+sage-core = { path = "../sage" }
